pyproject.toml
[project]
name = "gemini-fastapi"
version = "1.0.0"
description = "FastAPI Server built on Gemini Web API"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.129.0",
    "gemini-webapi>=1.19.2",
    "lmdb>=1.7.5",
    "loguru>=0.7.3",
    "orjson>=3.11.7",
    "pydantic-settings[yaml]>=2.12.0",
    "uvicorn>=0.40.0",
    "uvloop>=0.22.1; sys_platform != 'win32'",
    "google-generativeai (>=0.8.6,<0.9.0)",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "W", "I", "RUF"]
lint.ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
